Added Check Functionality + Typewriter Script
The act menu will now have monsters appear in it (similar to the fight button sans health bar).
The act menu will display all actions for the monster (currently only check).
Check will display the specified defence and attack of a target (the de and atk are seperate from the monster's stats so it can be replaced with a string for example)
A typewriter function has been added and is currently only used in obj_battle.