Improved bullet board manipulation
The battle board is now a separate object which can be rotated and such.

Going from player's turn to the monster's turn will now put the player in the centre of the bullet board immediately. The soul is also controllable during the Monster Speech state

Fixed issue with being able to back out of the first menu causing  the player to go to the bullet board.