Added functionality to ATK and DEF
Attack and Defence has been added with an invisible base of 10 for each.

Attack is based on the total of lv, item and base attacks  subtracted by enemy defence with a random of up to 2 additional points. This is multiplied by the accuracy bonus of the varying from 0 - 2 with attacks 12 pixels from the centre dealing 2.2 damage.

Defence is a reduction of damage an enemy inflicts. Damage to the player is calculated by the enemy ATK subtracted by the player defence.

Damage can not be reduced to below 1 for attacks.