Add defensive items and adjusted attacks
The faded ribbon and bandage have been added. When an armour item is equipped for the first time it will give Frisk the bandage, any other time it will give the previously equipped item.

Attacks now are calculated in the attack animation object instead of when the battle starts. Additional obj_battle now has t_atk and t_def variables that will be used to define temporary buffs for attack and defence respectively.