Added interaction and save points
Save points have been added which act as a way of testing interaction in the mean time. They do not save the game yet however they do heal the player.

Additionally, the player and save point have adjustable depth in order to always be in from of objects lower on the screen.