A fucking lot
Saving and loading is now controlled via a menu. Save points wills queue a save box in order to either save or return to the game.

Loading is controlled via the start menu. When the game loads with a save file intact it will transform into the flowey room like in regular Undertale else it will display the instructions page.

A name selection screen has been added which can be accessed by backing out of the use previous name selection. This is only done when a save is not present. If a save is present it disables entering a new name.

A settings screen has been added which is meant to be reminiscent of the joystick config/ console settings screen. This screen serves little function currently however the exit button and seasonal music have already been set.